Confirm OS updates are needed
Added the ability for BJPD to confirm OS updates are in fact still needed before prompting. This prevents BJPD from prompting if the user self-updated prior to the policy being run. To support this, also added parameter 7 to specify if it is being used for OS updates (default is no if not specified so should not break existing policies).